Refactor counters
Use fixed size arrays instead of hashmaps. This turns out to improve
runtime rather dramatically, to the point that no measurement is needed.

It looks slightly uglier but it is a lot faster.

Implement a daily counter that will be used to track daily changes in
cases to track daily changes in cases.